Multiple small fixes
Pawn moves and shipt moves and rings size are now configurable.
An exit routine was implemented which can be triggered with a button or
by the game end. Temp ini files are deleted on exit.
Payer order is shown correctly on init.
Digging, both the icon and the action, is disabled for pawns located on
a landscape with and empty resource stack.
The config struct is added to the game class for easy access.