Adjusted grid architecture and interface improve.
Divided Grid class in a Hexgrid class, which is general and can be used
for other games, and a Grid class, which inherits from Hexgrid and is
game specific.
Added mututal references of visualiser, grid and game objects to prevent
excessive passing of arguments.
Added a procedure for balancing the random landscape tile drawing. Land
is drawn before full random (including water) to prevent running out of
land tiles. Number of land tiles does not need to be specified in config
anymore but is derived from the number of tiles on the board.
Added message and log functions.
Added color coding to player names and landscape types in the interface.
Activated player is indicated in the points overview.