Game-breaking bugs and annoyances after playtest 1.
Bugfixes:
The move resource button is now located above the list of resources. The
list can still move out of reachable screen, but at least it will now
break the game.
Pawns can now unboard again. The problem was that unboarding is
conditional on pawns having moves. The moves for pawns were reset after
each turn by scanning the grid objects list, but pawns on boats are not
in that list. I replaced it with a list of object references for each
player which is scanned.

Improvements:
Clicking a boat (enemy or friendly) now shows whether there is a pawn
present.
Tiles next to harbours and within moveable range are now always accessible, also when not on the ring for pawns with ring reachability.
The tier 1 assignment requirements were halved.