Removed references to assignment buttons in game c
The assignment buttons were directly enabled/disabled in the game class.
This goes contrary to the split in visualiser and game and will later
cause problems. Fixed this by adding an enable/disable function for both
buttons in the visualiser class.
The layout of the popup menus had become messy, fixed this.